requirements.txt
streamlit==1.41.1
moviepy==2.1.2
ffmpeg-python==0.2.0
SpeechRecognition==3.8.1
textblob==0.17.1
nltk==3.6.7
matplotlib==3.5.1
plotly==5.5.0
pydantic==1.8.2